:bug: Fix broken favicon path.
The browser is making a request for the favicon by appending
favicon-32x32.png to the path that the user is at in the
application.

E.g. '/feedback/[id]/edit/favicon-32x32.png'

This is because the code is using a relative url instead of an
absolute url.

This change will fix that bug.